fix signature

+++ src/float_ieee754.h
@@ -2,5 +2,5 @@
 
 float u32_to_f32(uint32_t u);
 double u64_to_f64(uint64_t u);
-uint32_t f32_to_u32(double f);
+uint32_t f32_to_u32(float f);
 uint64_t f64_to_u64(double d);
